+++ injectable/example/lib/injector/injector_dev.config.dart
@@ -29,12 +29,18 @@ extension GetItInjectableX on _i174.GetIt {
       dispose: (i) => i.dispose(),
     );
     gh.singleton<_i978.ConstService>(() => const _i978.ConstService());
-    gh.factoryParam<_i978.IService, String?, dynamic>(
-      (param, _) => _i978.ServiceImpl(param),
-      instanceName: 'ServiceImpl',
-      registerFor: {_dev},
+    gh.factory<_i978.IService>(() => _i978.ServiceImpl());
+    gh.factoryCached<_i978.Model>(() => _i978.ModelX());
+    gh.factoryParam<_i978.LoggerService, String, dynamic>(
+      (name, _) => _i978.LoggerService(name),
     );
-    gh.factory<_i978.Model>(() => _i978.ModelX());
+    gh.factoryAsync<_i253.Repo>(
+      () => _i253.Repo.asyncRepo(gh<_i978.IService>()),
+    );
+    gh.singletonAsync<_i978.PostConstructableService>(() {
+      final i = _i978.PostConstructableService(gh<_i978.IService>());
+      return i.init().then((_) => i);
+    });
     await gh.factoryAsync<_i978.AbstractService>(
       () => _i978.AsyncService.create(
         gh<Set<String>>(instanceName: '__environments__'),
@@ -43,17 +49,14 @@ extension GetItInjectableX on _i174.GetIt {
       preResolve: true,
     );
     gh.lazySingletonAsync<_i253.Repo>(
-      () => registerModule.getRepo(
-        gh<_i978.IService>(instanceName: 'ServiceImpl'),
-      ),
+      () => registerModule.getRepo(gh<_i978.IService>()),
       instanceName: 'Repo',
-      registerFor: {_dev},
       dispose: _i253.disposeRepo,
     );
-    gh.singletonAsync<_i978.PostConstructableService>(() {
-      final i = _i978.PostConstructableService(gh<_i978.IService>());
-      return i.init().then((_) => i);
-    });
+    gh.factoryParam<_i978.ConfigurableService, String, String>(
+      (apiKey, baseUrl) =>
+          _i978.ConfigurableService(apiKey: apiKey, baseUrl: baseUrl),
+    );
     return this;
   }
 }