+++ injectable/example/lib/injector/injector_dev.dart
@@ -4,13 +4,10 @@ import 'package:injectable/injectable.dart';
 import 'injector_dev.config.dart';
 
 @InjectableInit(preferRelativeImports: false, generateForEnvironments: {dev})
-configDevInjector(
+Future<GetIt> configDevInjector(
   GetIt getIt, {
   String? env,
   EnvironmentFilter? environmentFilter,
 }) {
-  return getIt.init(
-    environmentFilter: environmentFilter,
-    environment: env,
-  );
+  return getIt.init(environmentFilter: environmentFilter, environment: env);
 }