+++ injectable_generator/lib/builder.dart
@@ -1,3 +1,4 @@
+// coverage:ignore-file
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';