+++ injectable_generator/lib/code_builder/library_builder.dart
@@ -3,6 +3,7 @@ import 'package:injectable_generator/code_builder/builder_utils.dart';
 import 'package:injectable_generator/models/dependency_config.dart';
 import 'package:injectable_generator/models/dispose_function_config.dart';
 import 'package:injectable_generator/models/external_module_config.dart';
+import 'package:injectable_generator/models/importable_type.dart';
 import 'package:injectable_generator/models/injected_dependency.dart';
 import 'package:injectable_generator/models/module_config.dart';
 import 'package:injectable_generator/utils.dart';
@@ -49,15 +50,11 @@ mixin SharedGeneratorCode {
     );
 
     final ref = typeRefer(dep.typeImpl, targetFile);
-    if (dep.constructorName?.isNotEmpty == true) {
+    if (dep.constructorName.isNotEmpty == true) {
       final constructor = dep.canBeConst
           ? ref.constInstanceNamed
           : ref.newInstanceNamed;
-      return constructor(
-        dep.constructorName!,
-        positionalParams,
-        namedParams,
-      );
+      return constructor(dep.constructorName, positionalParams, namedParams);
     } else {
       final constructor = dep.canBeConst ? ref.constInstance : ref.newInstance;
       return constructor(positionalParams, namedParams);
@@ -81,9 +78,7 @@ mixin SharedGeneratorCode {
         if (iDep.instanceName != null)
           'instanceName': literalString(iDep.instanceName!),
       },
-      [
-        typeRefer(iDep.type, targetFile, false),
-      ],
+      [typeRefer(iDep.type, targetFile, false)],
     );
     return isAsync ? expression.awaited : expression;
   }
@@ -102,11 +97,14 @@ class LibraryGenerator with SharedGeneratorCode {
   final Set<ExternalModuleConfig> microPackagesModulesBefore,
       microPackagesModulesAfter;
 
+  final bool generateAccessors;
+
   LibraryGenerator({
     required List<DependencyConfig> dependencies,
     required this.initializerName,
     this.targetFile,
     this.asExtension = false,
+    this.generateAccessors = false,
     this.microPackageName,
     this.microPackagesModulesBefore = const {},
     this.microPackagesModulesAfter = const {},
@@ -140,10 +138,9 @@ class LibraryGenerator with SharedGeneratorCode {
 
     final isMicroPackage = microPackageName != null;
 
-    throwIf(
-      isMicroPackage && scopedDeps.length > 1,
-      'Scopes are not supported in micro package modules!',
-    );
+    if (isMicroPackage && scopedDeps.length > 1) {
+      throw 'Scopes are not supported in micro package modules!';
+    }
 
     // make sure root scope is always generated even if empty
     if (!scopedDeps.containsKey(null)) {
@@ -154,11 +151,11 @@ class LibraryGenerator with SharedGeneratorCode {
       ...scopedBeforeExternalModules.keys,
       ...scopedAfterExternalModules.keys,
     };
-    final initMethods = <Method>[];
+    final extMethods = <Method>[];
     for (final scope in allScopeKeys) {
       final scopeDeps = scopedDeps[scope];
       final isRootScope = scope == null;
-      initMethods.add(
+      extMethods.add(
         InitMethodGenerator(
           scopeDependencies: scopeDeps ?? [],
           targetFile: targetFile,
@@ -178,56 +175,130 @@ class LibraryGenerator with SharedGeneratorCode {
       );
     }
 
+    generateAccessorMethods(extMethods);
+
     return Library(
       (b) => b
         ..comments.addAll([
           'ignore_for_file: type=lint',
           'coverage:ignore-file',
         ])
-        ..body.addAll(
-          [
-            ...environments.map(
-              (env) => Field(
-                (b) => b
-                  ..name = '_$env'
-                  ..type = refer('String')
-                  ..assignment = literalString(env).code
-                  ..modifier = FieldModifier.constant,
-              ),
+        ..body.addAll([
+          ...environments.map(
+            (env) => Field(
+              (b) => b
+                ..name = '_$env'
+                ..type = refer('String')
+                ..assignment = literalString(env).code
+                ..modifier = FieldModifier.constant,
             ),
-            if (!isMicroPackage) ...[
-              if (asExtension)
-                Extension(
-                  (b) => b
-                    ..name = 'GetItInjectableX'
-                    ..on = _getItRefer
-                    ..methods.addAll(initMethods),
-                )
-              else
-                ...initMethods,
-            ],
-
-            if (isMicroPackage)
-              Class(
+          ),
+          if (!isMicroPackage) ...[
+            if (asExtension)
+              Extension(
                 (b) => b
-                  ..name = '${capitalize(microPackageName!)}PackageModule'
-                  ..extend = refer(
-                    'MicroPackageModule',
-                    _injectableImport,
-                  )
-                  ..methods.add(initMethods.first),
-              ),
-
-            // build modules
-            ...modules.map(
-              (module) => _buildModule(
-                module,
-                dependencies.where((e) => e.moduleConfig == module),
-              ),
-            ),
+                  ..name = 'GetItInjectableX'
+                  ..on = _getItRefer
+                  ..methods.addAll(extMethods),
+              )
+            else
+              ...extMethods,
           ],
-        ),
+
+          if (isMicroPackage)
+            Class(
+              (b) => b
+                ..name = '${capitalize(microPackageName!)}PackageModule'
+                ..extend = refer('MicroPackageModule', _injectableImport)
+                ..methods.add(extMethods.first),
+            ),
+
+          // build modules
+          ...modules.map(
+            (module) => _buildModule(
+              module,
+              dependencies.where((e) => e.moduleConfig == module),
+            ),
+          ),
+        ]),
     );
+  }
+
+  void generateAccessorMethods(List<Method> extMethods) {
+    if (!generateAccessors) return;
+    final usedTypes = <ImportableType>{};
+    for (final dep in dependencies) {
+      if (dep.isFromModule || !usedTypes.add(dep.typeImpl)) {
+        continue;
+      }
+      final passesArgs =
+          dep.dependencies.any((d) => d.isFactoryParam) ||
+          dep.instanceName != null;
+      final isAsyncOrHasAsyncDep =
+          dep.isAsync ||
+          dep.dependencies.any(dependencies.isAsyncOrHasAsyncDependency);
+      final returns = isAsyncOrHasAsyncDep
+          ? TypeReference(
+              (b) => b
+                ..symbol = 'Future'
+                ..types.add(typeRefer(dep.typeImpl, targetFile)),
+            )
+          : typeRefer(dep.typeImpl, targetFile);
+      extMethods.add(
+        Method(
+          (b) {
+            b
+              ..name = toCamelCase(dep.typeImpl.name)
+              ..returns = returns
+              ..type = passesArgs ? null : MethodType.getter
+              ..lambda = true;
+            // add parameters for factory params
+            if (dep.instanceName != null) {
+              b.optionalParameters.add(
+                Parameter(
+                  (pb) => pb
+                    ..named = true
+                    ..name = 'instanceName'
+                    ..type = nullableRefer('String', nullable: true),
+                ),
+              );
+            }
+            for (final iDep in dep.dependencies.where(
+              (d) => d.isFactoryParam,
+            )) {
+              b.optionalParameters.add(
+                Parameter(
+                  (pb) => pb
+                    ..required = iDep.isRequired && !iDep.type.isNullable
+                    ..named = true
+                    ..name = iDep.paramName
+                    ..type = typeRefer(iDep.type, targetFile),
+                ),
+              );
+            }
+
+            int paramIndex = 0;
+            b.body =
+                refer(
+                      isAsyncOrHasAsyncDep ? 'getAsync' : 'get',
+                    )
+                    .call(
+                      [],
+                      {
+                        if (dep.instanceName != null)
+                          'instanceName': refer('instanceName'),
+                        for (final iDep in dep.dependencies.where(
+                          (d) => d.isFactoryParam,
+                        ))
+                          'param${++paramIndex}': refer(iDep.paramName),
+                      },
+                      [typeRefer(dep.typeImpl, targetFile)],
+                    )
+                    .code;
+          },
+        ),
+      );
+    }
   }
 
   Class _buildModule(ModuleConfig module, Iterable<DependencyConfig> deps) {
@@ -386,13 +457,11 @@ class InitMethodGenerator with SharedGeneratorCode {
           : _getItRefer;
     }
 
-    final ghBuilder = refer('GetItHelper', _injectableImport).newInstance(
-      [
-        getInstanceRefer,
-        refer('environment'),
-        refer('environmentFilter'),
-      ],
-    );
+    final ghBuilder = refer('GetItHelper', _injectableImport).newInstance([
+      getInstanceRefer,
+      refer('environment'),
+      refer('environmentFilter'),
+    ]);
 
     return Method(
       (b) => b
@@ -423,10 +492,7 @@ class InitMethodGenerator with SharedGeneratorCode {
               (b) => b
                 ..named = true
                 ..name = 'environment'
-                ..type = nullableRefer(
-                  'String',
-                  nullable: true,
-                ),
+                ..type = nullableRefer('String', nullable: true),
             ),
             Parameter(
               (b) => b
@@ -443,10 +509,7 @@ class InitMethodGenerator with SharedGeneratorCode {
                 (b) => b
                   ..named = true
                   ..name = 'constructorCallback'
-                  ..type = nullableRefer(
-                    'T Function<T>(T)',
-                    nullable: true,
-                  ),
+                  ..type = nullableRefer('T Function<T>(T)', nullable: true),
               ),
           ] else if (!isMicroPackage)
             Parameter(
@@ -526,30 +589,41 @@ class InitMethodGenerator with SharedGeneratorCode {
     if (dep.injectableType == InjectableType.factory) {
       final hasFactoryParams = dep.dependencies.any((d) => d.isFactoryParam);
       if (hasFactoryParams) {
-        funcReferName = isOrHasAsyncDep ? 'factoryParamAsync' : 'factoryParam';
+        funcReferName = switch ((isOrHasAsyncDep, dep.cache == true)) {
+          (true, true) => 'factoryCachedParamAsync',
+          (false, true) => 'factoryCachedParam',
+          (false, false) => 'factoryParam',
+          (true, false) => 'factoryParamAsync',
+        };
+
         factoryParams.addAll(_resolveFactoryParams(dep));
       } else {
-        funcReferName = isOrHasAsyncDep ? 'factoryAsync' : 'factory';
+        funcReferName = switch ((isOrHasAsyncDep, dep.cache == true)) {
+          (true, true) => 'factoryCachedAsync',
+          (false, true) => 'factoryCached',
+          (false, false) => 'factory',
+          (true, false) => 'factoryAsync',
+        };
       }
     } else if (dep.injectableType == InjectableType.lazySingleton) {
       funcReferName = isOrHasAsyncDep ? 'lazySingletonAsync' : 'lazySingleton';
     }
-    throwIf(funcReferName == null, 'Injectable type is not supported');
+    if (funcReferName == null) {
+      throw 'Injectable type is not supported';
+    }
 
     final instanceBuilder = dep.isFromModule
         ? _buildInstanceForModule(dep)
         : _buildInstance(dep);
     final instanceBuilderCode = _buildInstanceBuilderCode(instanceBuilder, dep);
-    final registerExpression = _ghLocalRefer.property(funcReferName!).call(
+    final registerExpression = _ghLocalRefer.property(funcReferName).call(
       [
         Method(
           (b) => b
             ..lambda = instanceBuilderCode is! Block
             ..modifier = hasAsyncDep ? MethodModifier.async : null
             ..requiredParameters.addAll(
-              factoryParams.keys.map(
-                (name) => Parameter((b) => b.name = name),
-              ),
+              factoryParams.keys.map((name) => Parameter((b) => b.name = name)),
             )
             ..body = instanceBuilderCode,
         ).closure,
@@ -558,9 +632,7 @@ class InitMethodGenerator with SharedGeneratorCode {
         if (dep.instanceName != null)
           'instanceName': literalString(dep.instanceName!),
         if (dep.environments.isNotEmpty == true)
-          'registerFor': literalSet(
-            dep.environments.map((e) => refer('_$e')),
-          ),
+          'registerFor': literalSet(dep.environments.map((e) => refer('_$e'))),
         if (dep.preResolve == true) 'preResolve': literalBool(true),
         if (dep.disposeFunction != null)
           'dispose': _getDisposeFunctionAssignment(dep.disposeFunction!),
@@ -664,14 +736,10 @@ class InitMethodGenerator with SharedGeneratorCode {
           'instanceName': literalString(dep.instanceName!),
         if (dep.dependsOn.isNotEmpty)
           'dependsOn': literalList(
-            dep.dependsOn.map(
-              (e) => typeRefer(e, targetFile),
-            ),
+            dep.dependsOn.map((e) => typeRefer(e, targetFile)),
           ),
         if (dep.environments.isNotEmpty)
-          'registerFor': literalSet(
-            dep.environments.map((e) => refer('_$e')),
-          ),
+          'registerFor': literalSet(dep.environments.map((e) => refer('_$e'))),
         if (dep.signalsReady != null)
           'signalsReady': literalBool(dep.signalsReady!),
         if (dep.preResolve == true) 'preResolve': literalBool(true),
@@ -696,15 +764,10 @@ class InitMethodGenerator with SharedGeneratorCode {
 
     return refer(toCamelCase(module.type.name)).newInstanceNamed(
       module.initializerName,
-      dep.positionalDependencies.map(
-        (iDep) => _buildParamAssignment(iDep),
-      ),
+      dep.positionalDependencies.map((iDep) => _buildParamAssignment(iDep)),
       Map.fromEntries(
         dep.namedDependencies.map(
-          (iDep) => MapEntry(
-            iDep.paramName,
-            _buildParamAssignment(iDep),
-          ),
+          (iDep) => MapEntry(iDep.paramName, _buildParamAssignment(iDep)),
         ),
       ),
     );
@@ -728,7 +791,5 @@ class InitMethodGenerator with SharedGeneratorCode {
 bool moduleHasOverrides(Iterable<DependencyConfig> deps) {
   return deps
       .where((d) => d.moduleConfig?.isAbstract == true)
-      .any(
-        (d) => d.dependencies.isNotEmpty == true,
-      );
+      .any((d) => d.dependencies.isNotEmpty == true);
 }