+++ injectable_generator/lib/generators/injectable_generator.dart
@@ -7,17 +7,12 @@ import 'package:injectable/injectable.dart';
 import 'package:injectable_generator/models/dependency_config.dart';
 import 'package:injectable_generator/resolvers/dependency_resolver.dart';
 import 'package:injectable_generator/resolvers/importable_type_resolver.dart';
-import 'package:injectable_generator/utils.dart';
 import 'package:source_gen/source_gen.dart';
 
-const TypeChecker _typeChecker = TypeChecker.typeNamed(
-  Injectable,
-  inPackage: 'injectable',
-);
-const TypeChecker _moduleChecker = TypeChecker.typeNamed(
-  Module,
-  inPackage: 'injectable',
-);
+import '../resolvers/utils.dart';
+
+const _typeChecker = TypeChecker.typeNamed(Injectable, inPackage: 'injectable');
+const _moduleChecker = TypeChecker.typeNamed(Module, inPackage: 'injectable');
 
 class InjectableGenerator implements Generator {
   RegExp? _classNameMatcher, _fileNameMatcher;
@@ -38,6 +33,8 @@ class InjectableGenerator implements Generator {
   @override
   FutureOr<String?> generate(LibraryReader library, BuildStep buildStep) async {
     final allDepsInStep = <DependencyConfig>[];
+    final libs = await buildStep.resolver.libraries.toList();
+    final resolver = getResolver(libs);
     for (var clazz in library.classes) {
       if (_moduleChecker.hasAnnotationOfExact(clazz)) {
         throwIf(
@@ -52,17 +49,13 @@ class InjectableGenerator implements Generator {
         for (var element in executables) {
           if (element.isPrivate) continue;
           allDepsInStep.add(
-            DependencyResolver(
-              getResolver(await buildStep.resolver.libraries.toList()),
-            ).resolveModuleMember(clazz, element),
+            DependencyResolver(resolver).resolveModuleMember(clazz, element),
           );
         }
       } else if (_hasInjectable(clazz) ||
           (_autoRegister && _hasConventionalMatch(clazz))) {
         allDepsInStep.add(
-          DependencyResolver(
-            getResolver(await buildStep.resolver.libraries.toList()),
-          ).resolve(clazz),
+          DependencyResolver(resolver).resolve(clazz),
         );
       }
     }