+++ injectable_generator/lib/models/dependency_config.dart
@@ -20,7 +20,7 @@ class DependencyConfig {
   final String? instanceName;
   final bool? signalsReady;
   final List<String> environments;
-  final String? constructorName;
+  final String constructorName;
   final String? postConstruct;
   final bool isAsync;
   final bool postConstructReturnsSelf;
@@ -31,6 +31,7 @@ class DependencyConfig {
   final DisposeFunctionConfig? disposeFunction;
   final int orderPosition;
   final String? scope;
+  final bool cache;
 
   DependencyConfig({
     required this.type,
@@ -51,6 +52,7 @@ class DependencyConfig {
     this.scope,
     this.postConstructReturnsSelf = false,
     this.postConstruct,
+    this.cache = false,
   });
 
   // used for testing
@@ -60,12 +62,14 @@ class DependencyConfig {
     List<String> deps = const [],
     List<String> envs = const [],
     int order = 0,
+    bool cache = false,
   }) {
     return DependencyConfig(
-      type: ImportableType(name: type, import: type),
+      type: ImportableType(name: type),
       typeImpl: ImportableType(name: typeImpl ?? type),
       environments: envs,
       orderPosition: order,
+      cache: cache,
       dependencies: deps
           .map(
             (e) => InjectedDependency(
@@ -94,6 +98,7 @@ class DependencyConfig {
           : InjectableType.singleton,
       environments: envs,
       orderPosition: order,
+      cache: false,
       dependencies: deps
           .map(
             (e) => InjectedDependency(
@@ -131,6 +136,7 @@ class DependencyConfig {
           disposeFunction == other.disposeFunction &&
           scope == other.scope &&
           moduleConfig == other.moduleConfig &&
+          cache == other.cache &&
           postConstruct == other.postConstruct &&
           postConstructReturnsSelf == other.postConstructReturnsSelf &&
           orderPosition == other.orderPosition);
@@ -153,6 +159,7 @@ class DependencyConfig {
       canBeConst.hashCode ^
       orderPosition.hashCode ^
       postConstruct.hashCode ^
+      cache.hashCode ^
       postConstructReturnsSelf.hashCode ^
       scope.hashCode;
 
@@ -201,6 +208,7 @@ class DependencyConfig {
       injectableType: json['injectableType'],
       instanceName: json['instanceName'],
       signalsReady: json['signalsReady'],
+      cache: (json['cache'] as bool?) ?? false,
       environments: json['environments']?.cast<String>(),
       constructorName: json['constructorName'],
       postConstruct: json['postConstruct'],
@@ -230,8 +238,9 @@ class DependencyConfig {
     "environments": environments,
     "dependencies": dependencies.map((v) => v.toJson()).toList(),
     if (instanceName != null) "instanceName": instanceName,
+    "cache": cache,
     if (signalsReady != null) "signalsReady": signalsReady,
-    if (constructorName != null) "constructorName": constructorName,
+    "constructorName": constructorName,
     if (postConstruct != null) "postConstruct": postConstruct,
     "orderPosition": orderPosition,
     if (scope != null) "scope": scope,