+++ injectable_generator/lib/resolvers/dependency_resolver.dart
@@ -6,57 +6,54 @@ import 'package:injectable_generator/models/dispose_function_config.dart';
 import 'package:injectable_generator/models/importable_type.dart';
 import 'package:injectable_generator/models/injected_dependency.dart';
 import 'package:injectable_generator/models/module_config.dart';
+import 'package:injectable_generator/resolvers/utils.dart';
 import 'package:injectable_generator/utils.dart';
 import 'package:source_gen/source_gen.dart';
 
 import '../injectable_types.dart';
 import 'importable_type_resolver.dart';
 
-const TypeChecker _namedChecker = TypeChecker.typeNamed(
-  Named,
-  inPackage: 'injectable',
-);
-const TypeChecker _ignoredChecker = TypeChecker.typeNamed(
-  IgnoreParam,
-  inPackage: 'injectable',
-);
-const TypeChecker _injectableChecker = TypeChecker.typeNamed(
+const _injectableChecker = TypeChecker.typeNamed(
   Injectable,
   inPackage: 'injectable',
 );
-const TypeChecker _envChecker = TypeChecker.typeNamed(
-  Environment,
+const _lazySingletonChecker = TypeChecker.typeNamed(
+  LazySingleton,
   inPackage: 'injectable',
 );
-const TypeChecker _preResolveChecker = TypeChecker.typeNamed(
-  PreResolve,
-  inPackage: 'injectable',
-);
-const TypeChecker _factoryParamChecker = TypeChecker.typeNamed(
-  FactoryParam,
-  inPackage: 'injectable',
-);
-const TypeChecker _scopeChecker = TypeChecker.typeNamed(
-  Scope,
-  inPackage: 'injectable',
-);
-const TypeChecker _factoryMethodChecker = TypeChecker.typeNamed(
-  FactoryMethod,
-  inPackage: 'injectable',
-);
-const TypeChecker _disposeMethodChecker = TypeChecker.typeNamed(
-  DisposeMethod,
-  inPackage: 'injectable',
-);
-const TypeChecker _postConstructChecker = TypeChecker.typeNamed(
-  PostConstruct,
+const _singletonChecker = TypeChecker.typeNamed(
+  Singleton,
   inPackage: 'injectable',
 );
 
-const TypeChecker _orderChecker = TypeChecker.typeNamed(
-  Order,
+const _namedChecker = TypeChecker.typeNamed(Named, inPackage: 'injectable');
+const _ignoredChecker = TypeChecker.typeNamed(
+  IgnoreParam,
   inPackage: 'injectable',
 );
+const _envChecker = TypeChecker.typeNamed(Environment, inPackage: 'injectable');
+const _preResolveChecker = TypeChecker.typeNamed(
+  PreResolve,
+  inPackage: 'injectable',
+);
+const _factoryParamChecker = TypeChecker.typeNamed(
+  FactoryParam,
+  inPackage: 'injectable',
+);
+const _scopeChecker = TypeChecker.typeNamed(Scope, inPackage: 'injectable');
+const _factoryMethodChecker = TypeChecker.typeNamed(
+  FactoryMethod,
+  inPackage: 'injectable',
+);
+const _disposeMethodChecker = TypeChecker.typeNamed(
+  DisposeMethod,
+  inPackage: 'injectable',
+);
+const _postConstructChecker = TypeChecker.typeNamed(
+  PostConstruct,
+  inPackage: 'injectable',
+);
+const _orderChecker = TypeChecker.typeNamed(Order, inPackage: 'injectable');
 
 class DependencyResolver {
   final ImportableTypeResolver _typeResolver;
@@ -71,12 +68,13 @@ class DependencyResolver {
   String? _instanceName;
   bool _isAsync = false;
   bool _canBeConst = false;
-  String? _constructorName;
+  String _constructorName = '';
   final List<InjectedDependency> _dependencies = [];
   ModuleConfig? _moduleConfig;
   DisposeFunctionConfig? _disposeFunctionConfig;
   int? _order;
   String? _scope;
+  bool _cache = false;
 
   DependencyResolver(this._typeResolver);
 
@@ -139,23 +137,20 @@ class DependencyResolver {
       annotatedElement,
       throwOnUnresolved: false,
     );
-
     DartType? abstractType;
     ExecutableElement? disposeFuncFromAnnotation;
     List<String>? inlineEnv;
     if (injectableAnnotation != null) {
       final injectable = ConstantReader(injectableAnnotation);
-      if (injectable.instanceOf(
-        TypeChecker.typeNamed(LazySingleton, inPackage: 'injectable'),
-      )) {
+      _cache = injectable.peek('cache')?.boolValue == true;
+
+      if (injectable.instanceOf(_lazySingletonChecker)) {
         _injectableType = InjectableType.lazySingleton;
         disposeFuncFromAnnotation = injectable
             .peek('dispose')
             ?.objectValue
             .toFunctionValue();
-      } else if (injectable.instanceOf(
-        TypeChecker.typeNamed(Singleton, inPackage: 'injectable'),
-      )) {
+      } else if (injectable.instanceOf(_singletonChecker)) {
         _injectableType = InjectableType.singleton;
         _signalsReady = injectable.peek('signalsReady')?.boolValue;
         disposeFuncFromAnnotation = injectable
@@ -195,9 +190,14 @@ class DependencyResolver {
         '[${clazz.displayName}] is not a subtype of [${abstractType.nameWithoutSuffix}]',
         element: clazz,
       );
-
       _type = _typeResolver.resolveType(abstractSubtype!);
     }
+
+    throwIf(
+      _cache == true && _injectableType != InjectableType.factory,
+      'Only factory types can be cached',
+      element: clazz,
+    );
 
     _environments =
         inlineEnv ??
@@ -363,6 +363,7 @@ class DependencyResolver {
           isFactoryParam: isFactoryParam,
           paramName: param.displayName,
           isPositional: param.isPositional,
+          isRequired: param.isRequired,
         ),
       );
     }
@@ -465,6 +466,7 @@ class DependencyResolver {
       scope: _scope,
       postConstruct: postConstruct,
       postConstructReturnsSelf: postConstructReturnsSelf,
+      cache: _cache,
     );
   }
 }